Move concatenation tests to their own file
